Solve incorrect lookup count
Previous method would not count all lookups of includes that did not have IP addresses of their own. If an included subspf.domain.com only contains other includes and no ip4: or ip6: directives, then the domain subspf.domain.com does not appear in the SPFSourceDomain property in the ReturnValues and is thus not counted towards the total.

New method instead combines the Referrer and SPFSourceDomain properties to create an exhaustive array of all used lookups.

It now also excludes the original domain lookup that is not counted towards the lookup limit.